requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
mlflow==3.1.4
passlib[bcrypt]==1.7.4  
PyJWT==2.9.0  
bcrypt==4.1.2  
